Add form for adding new sets
This commit adds a form and API endpoint for creating new sets, which is accessible to
only admin users at `/sets/new` and an authorized POST request to `/sets`
respectively. This may be considered something of a minimum viable product in this commit. The code could most
certainly be improved and cleaned up, but there are so many features
required to make this feature-complete that I don't think it's worth
fully implementing at this point. A non-exhaustive list of features that
have been entirely neglected from this commit:

* Removing cases from a set while adding it
* Editing a set later
* Editing an alg later
* Associating an existing case with a new set
* A properly working form that validates on the frontend
* Anything to do with methods

As you can tell, all these features would keep someone busy for a while.
Such an early-stage project doesn't need all those at this point, and
it's probably a better use of time to wait to implement those when there
is a bit more data in the system so that there is actually something
that we can use when creating those features.