Fix sidebar showing duplicate set names
Some sets share the same name, because they refer to equivalent steps on
different puzzles. This commit adds the puzzle name to the link to the
set in the navbar.

Long-term another solution will need to be found, as right now the
navbar takes all the sets present in the database and creates an entry
for them, so it will probably be too big in the future. But for now,
this will work fine.